switch back from bfloat16

+++ python/train.py
@@ -230,7 +230,7 @@ def train_one_batch(y, state, loss_fn, condition_vec=None, use_amp=False, ctrl_f
     # Compute loss and do backwards pass
     if use_amp:
         with timer.section("forward"):
-            with torch.amp.autocast_mode.autocast(DEVICE.type, dtype=torch.bfloat16):
+            with torch.amp.autocast_mode.autocast(DEVICE.type):
                 loss, base_loss, *_ = loss_fn(y, state.model, condition_vec=condition_vec, ctrl=ctrl)
         with timer.section("backward"):
             state.scaler.scale(loss).backward()
